Move main API for interacting with high-level data objects from DataCollection to Data
