+++ glue/core/data_collection.py
@@ -410,138 +410,3 @@ class DataCollection(HubListener):
 
     def __nonzero__(self):
         return True
-
-    def get_object(self, data_label, cls=None):
-        """
-        Get data represented as a non-glue object, using the translation
-        machinery.
-
-        Parameters
-        ----------
-        data_label : str
-            The name or index of the dataset to retrieve.
-        cls : type, optional
-            The class to use for representing the data object. If a non-glue
-            data object was added to the data collection, it should automatically
-            be returned using the same class as it was provided in, and this
-            argument shouldn't be needed.
-        """
-
-        if len(self) == 0:
-            raise ValueError("No datasets are present in the data collection")
-
-        if data_label is None:
-            if len(self) == 1:
-                data = self[0]
-            else:
-                raise ValueError("Multiple datasets are present in the data collection, "
-                                 "use the data_label argument to specify which one to retrieve.")
-        else:
-            data = self[data_label]
-
-        if cls is None:
-            if hasattr(data, '_preferred_translation'):
-                cls = data._preferred_translation
-            else:
-                raise ValueError('Specify the object class to use with cls=')
-
-        handler, _ = data_translator.get_handler_for(cls)
-
-        return handler.to_object(data)
-
-    def get_subset_object(self, data_label=None, subset_id=None, cls=None):
-        """
-        Get subset represented as a non-glue object, using the translation
-        machinery.
-
-        Parameters
-        ----------
-        data_label : str
-            The name or index of the dataset to retrieve.
-        subset_id : str or int, optional
-            The name or index of the subset to retrieve.
-        cls : type, optional
-            The class to use for representing the data object. If a non-glue
-            data object was added to the data collection, it should automatically
-            be returned using the same class as it was provided in, and this
-            argument shouldn't be needed.
-        """
-
-        if len(self) == 0:
-            raise ValueError("No datasets are present in the data collection")
-
-        if data_label is None:
-            if len(self) == 1:
-                data = self[0]
-            else:
-                raise ValueError("Multiple datasets are present in the data collection, "
-                                 "use the data_label argument to specify which one to retrieve.")
-        else:
-            data = self[data_label]
-
-        if cls is None:
-            if hasattr(data, '_preferred_translation'):
-                cls = data._preferred_translation
-            else:
-                raise ValueError('Specify the object class to use with cls=')
-
-        if subset_id is None and len(data.subsets) == 1:
-            subset = data.subsets[0]
-        elif isinstance(subset_id, str):
-            matches = [subset for subset in data.subsets if subset.label == subset_id]
-            if len(matches) == 0:
-                raise ValueError("No subset found with the label '{0}'".format(subset_id))
-            elif len(matches) > 1:
-                raise ValueError("Several subsets were found with the label '{0}', use a numerical index instead".format(subset_id))
-            else:
-                subset = matches[0]
-        else:
-            subset = data.subsets[subset_id]
-
-        handler, _ = data_translator.get_handler_for(cls)
-
-        return handler.to_object(subset)
-
-    def get_selection_definition(self, data_label=None, subset_id=None, format=None):
-        """
-        Get subset state represented as a non-glue object, using the translation
-        machinery.
-
-        Parameters
-        ----------
-        data_label : str, optional
-            The name or index of the dataset to retrieve the subset for.
-        subset_id : str or int, optional
-            The name or index of the subset to retrieve.
-        format : str, optional
-            The format to translate the subset state to.
-        """
-
-        if len(self) == 0:
-            raise ValueError("No datasets are present in the data collection")
-
-        if data_label is None:
-            if len(self) == 1:
-                data = self[0]
-            else:
-                raise ValueError("Multiple datasets are present in the data collection, "
-                                 "use the data_label argument to specify which one to retrieve.")
-        else:
-            data = self[data_label]
-
-        if subset_id is None and len(data.subsets) == 1:
-            subset = data.subsets[0]
-        elif isinstance(subset_id, str):
-            matches = [subset for subset in self.subset_groups if subset.label == subset_id]
-            if len(matches) == 0:
-                raise ValueError("No subset found with the label '{0}'".format(subset_id))
-            elif len(matches) > 1:
-                raise ValueError("Several subsets were found with the label '{0}', use a numerical index instead".format(subset_id))
-            else:
-                subset = matches[0]
-        else:
-            subset = data.subsets[subset_id]
-
-        handler = subset_state_translator.get_handler_for(format)
-
-        return handler.to_object(subset)
